Fix for pip>21.1.3
Pip 21.2 onwards use _ProcessedDists class to contain all the metadata of a
package instead of a dictionary. This fix makes the necessary changes to account
for this.

A new function to compare pip versions is added to segment pippel.py script
according to the current environment's pip version.

Several typing hints and comments are added for clarity.

Signed-off-by: Arif Er <arifer612@pm.me>