Fix I/O logging error when (un)installing packages
The issue seems to lie in the usage of the `main()` command for InstallCommand
and UninstallCommand. The easiest and cleanest way out of it is to ensure that
the process is being called properly using `subprocess.check_call()` instead.

The `traceback` library is added to include more verbosity for error debugging
since just declaring the type of Exception raised does not shed enough light
onto the issue.

Signed-off-by: Arif Er <arifer612@pm.me>